+++ bin/commands/changes.rb
@@ -1,0 +1,19 @@
+# frozen_string_literal: true
+
+desc 'Display the changelog'
+command %i[changes changelog] do |c|
+  c.action do |_, _, _|
+    changelog = File.expand_path(File.join(File.dirname(__FILE__), '..', 'CHANGELOG.md'))
+    pagers = [
+      'mdless',
+      'mdcat',
+      'bat',
+      ENV['PAGER'],
+      'less -FXr',
+      ENV['GIT_PAGER'],
+      'more -r'
+    ]
+    pager = pagers.find { |cmd| TTY::Which.exist?(cmd.split.first) }
+    system %(#{pager} "#{changelog}")
+  end
+end